+++ tests/testthat/test-url.R
@@ -93,27 +93,67 @@ test_that("path always starts with /", {
   expect_equal(url_modify("https://x.com/abc", path = ""), "https://x.com/")
 })
 
+# modify query -------------------------------------------------------------
+
+test_that("can add, modify, and delete query components", {
+  expect_equal(
+    url_modify_query("http://test/path", new = "value"),
+    "http://test/path?new=value"
+  )
+  expect_equal(
+    url_modify_query("http://test/path", new = "one", new = "two"),
+    "http://test/path?new=one&new=two"
+  )
+  expect_equal(
+    url_modify_query("http://test/path?a=old&b=old", a = "new"),
+    "http://test/path?b=old&a=new"
+  )
+  expect_equal(
+    url_modify_query("http://test/path?remove=me&keep=this", remove = NULL),
+    "http://test/path?keep=this"
+  )
+})
+
+test_that("is idempotent", {
+  string <- "http://example.com/"
+  url <- url_parse(string)
+
+  expect_equal(url_modify_query(string), string)
+  expect_equal(url_modify_query(url), url)
+})
+
+test_that("validates inputs", {
+  url <- "http://example.com/"
+
+  expect_snapshot(error = TRUE, {
+    url_modify_query(1)
+    url_modify_query(url, 1)
+    url_modify_query(url, x = 1:2)
+  })
+})
+
+
 # query -------------------------------------------------------------------
 
 test_that("missing query values become empty strings", {
-  expect_equal(query_parse("?q="), list(q = ""))
-  expect_equal(query_parse("?q"), list(q = ""))
-  expect_equal(query_parse("?a&q"), list(a = "", q = ""))
+  expect_equal(url_query_parse("?q="), list(q = ""))
+  expect_equal(url_query_parse("?q"), list(q = ""))
+  expect_equal(url_query_parse("?a&q"), list(a = "", q = ""))
 })
 
 test_that("handles equals in values", {
-  expect_equal(query_parse("?x==&y=="), list(x = "=", y = "="))
+  expect_equal(url_query_parse("?x==&y=="), list(x = "=", y = "="))
  })
 
 test_that("empty queries become NULL", {
-  expect_equal(query_parse("?"), NULL)
-  expect_equal(query_parse(""), NULL)
+  expect_equal(url_query_parse("?"), NULL)
+  expect_equal(url_query_parse(""), NULL)
 })
 
 test_that("validates inputs", {
   expect_snapshot(error = TRUE, {
-    query_build(1:3)
-    query_build(list(x = 1:2, y = 1:3))
+    url_query_build(1:3)
+    url_query_build(list(x = 1:2, y = 1:3))
   })
 })